pyproject.toml
[project]
name = "ezmsg-unicorn"
version = "0.2.0"
description = "g.tec Unicorn integration for ezmsg"
authors = [
    { name = "Griffin Milsap", email = "griffin.milsap@gmail.com" },
    { name = "Miguel Wimbish", email = "miguel.wimbish@jhuapl.edu" },
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "ezmsg>=3.5.0",
    "pyqt5>=5.15.11; platform_system != 'Linux'",
]

[project.optional-dependencies]
dashboard = [
    "ezmsg-panel>=0.5.0",
    "ezmsg-sigproc>=1.4.2",
    "numpy<2.0.0",
    "vqf>=2.0.0",
]

[project.scripts]
dashboard = "ezmsg.unicorn.dashboard:dashboard"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ezmsg"]
